SWs working

+++ src/include/catalog/pipeline_query_fn.h
@@ -45,7 +45,7 @@ typedef struct ContQuery
 	Oid seqrelid;
 	int sw_step_factor;
 	int sw_step_ms;
-	int sw_interval_ms;
+	uint64 sw_interval_ms;
 	bool is_sw;
 
 	/* for transform */
